selfstudy/장고연습/practice12/requirements.txt
asgiref==3.5.0
Django==3.2
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
